+++ neo/sys/win32/SDL_win32_main.c
@@ -160,154 +160,6 @@ static void cleanup(void)
 	SDL_Quit();
 }
 
-/* Remove the output files if there was no output written */
-static void cleanup_output(void) {
-	FILE *file;
-	int empty;
-
-	/* Flush the output in case anything is queued */
-	fclose(stdout);
-	fclose(stderr);
-
-	/* Without redirection we're done */
-	if (!stdioRedirectEnabled) {
-		return;
-	}
-
-	/* See if the files have any output in them */
-	if ( stdoutPath[0] ) {
-		file = fopen(stdoutPath, TEXT("rb"));
-		if ( file ) {
-			empty = (fgetc(file) == EOF) ? 1 : 0;
-			fclose(file);
-			if ( empty ) {
-				remove(stdoutPath);
-			}
-		}
-	}
-	if ( stderrPath[0] ) {
-		file = fopen(stderrPath, TEXT("rb"));
-		if ( file ) {
-			empty = (fgetc(file) == EOF) ? 1 : 0;
-			fclose(file);
-			if ( empty ) {
-				remove(stderrPath);
-			}
-		}
-	}
-}
-
-extern int Win_GetHomeDir(char *dst, size_t size);
-
-/* Redirect the output (stdout and stderr) to a file */
-static void redirect_output(void)
-{
-#ifdef _WIN32_WCE
-	wchar_t path[MAX_PATH];
-#error "adapt homedir code for wchar_t!"
-#else
-	char path[MAX_PATH];
-	struct _stat st;
-
-	/* DG: use "My Documents/My Games/dhewm3" to write stdout.txt and stderr.txt
-	 *     instead of the binary, which might not be writable */
-	Win_GetHomeDir(path, sizeof(path));
-
-	if (_stat(path, &st) == -1) {
-		/* oops, "My Documents/My Games/dhewm3" doesn't exist - does My Games/ at least exist? */
-		char myGamesPath[MAX_PATH];
-		char* lastslash;
-		memcpy(myGamesPath, path, MAX_PATH);
-		lastslash = strrchr(myGamesPath, '/');
-		if (lastslash != NULL) {
-			*lastslash = '\0';
-		}
-		if (_stat(myGamesPath, &st) == -1) {
-			/* if My Documents/My Games/ doesn't exist, create it */
-			_mkdir(myGamesPath);
-		}
-		
-		_mkdir(path); /* create My Documents/My Games/dhewm3/ */
-	}
-	
-
-#endif
-	FILE *newfp;
-
-#if 0 /* DG: don't do this anymore. */
-	DWORD pathlen;
-	pathlen = GetModuleFileName(NULL, path, SDL_arraysize(path));
-	while ( pathlen > 0 && path[pathlen] != '\\' ) {
-		--pathlen;
-	}
-	path[pathlen] = '\0';
-#endif
-
-#ifdef _WIN32_WCE
-	wcsncpy( stdoutPath, path, SDL_arraysize(stdoutPath) );
-	wcsncat( stdoutPath, DIR_SEPERATOR STDOUT_FILE, SDL_arraysize(stdoutPath) );
-#else
-	SDL_strlcpy( stdoutPath, path, SDL_arraysize(stdoutPath) );
-	SDL_strlcat( stdoutPath, DIR_SEPERATOR STDOUT_FILE, SDL_arraysize(stdoutPath) );
-#endif
-
-	{ /* DG: rename old stdout log */
-#ifdef _WIN32_WCE
-		wchar_t stdoutPathBK[MAX_PATH];
-		wcsncpy( stdoutPathBK, path, SDL_arraysize(stdoutPath) );
-		wcsncat( stdoutPathBK, DIR_SEPERATOR TEXT("dhewm3log-old.txt"), SDL_arraysize(stdoutPath) );
-		_wrename( stdoutPath, stdoutpathBK );
-#else
-		char stdoutPathBK[MAX_PATH];
-		SDL_strlcpy( stdoutPathBK, path, SDL_arraysize(stdoutPath) );
-		SDL_strlcat( stdoutPathBK, DIR_SEPERATOR TEXT("dhewm3log-old.txt"), SDL_arraysize(stdoutPath) );
-		rename( stdoutPath, stdoutPathBK );
-#endif
-	} /* DG end */
-
-	/* Redirect standard input and standard output */
-	newfp = freopen(stdoutPath, TEXT("w"), stdout);
-
-#ifndef _WIN32_WCE
-	if ( newfp == NULL ) {	/* This happens on NT */
-#if !defined(stdout)
-		stdout = fopen(stdoutPath, TEXT("w"));
-#else
-		newfp = fopen(stdoutPath, TEXT("w"));
-		if ( newfp ) {
-			*stdout = *newfp;
-		}
-#endif
-	}
-#endif /* _WIN32_WCE */
-
-#ifdef _WIN32_WCE
-	wcsncpy( stderrPath, path, SDL_arraysize(stdoutPath) );
-	wcsncat( stderrPath, DIR_SEPERATOR STDOUT_FILE, SDL_arraysize(stdoutPath) );
-#else
-	SDL_strlcpy( stderrPath, path, SDL_arraysize(stderrPath) );
-	SDL_strlcat( stderrPath, DIR_SEPERATOR STDERR_FILE, SDL_arraysize(stderrPath) );
-#endif
-
-	newfp = freopen(stderrPath, TEXT("w"), stderr);
-#ifndef _WIN32_WCE
-	if ( newfp == NULL ) {	/* This happens on NT */
-#if !defined(stderr)
-		stderr = fopen(stderrPath, TEXT("w"));
-#else
-		newfp = fopen(stderrPath, TEXT("w"));
-		if ( newfp ) {
-			*stderr = *newfp;
-		}
-#endif
-	}
-#endif /* _WIN32_WCE */
-
-	setvbuf(stdout, NULL, _IOLBF, BUFSIZ);	/* Line buffered */
-	setbuf(stderr, NULL);			/* No buffering */
-	stdioRedirectEnabled = 1;
-}
-
 #if defined(_MSC_VER) && !defined(_WIN32_WCE)
 /* The VC++ compiler needs main defined */
 #define console_main main
@@ -397,6 +249,7 @@ int WINAPI WinMain(HINSTANCE hInst, HINSTANCE hPrev, LPSTR szCmdLine, int sw)
 		FreeLibrary(handle);
 	}
 
+#if 0 // DG: output redirection is now done in dhewm3's main() aka SDL_main()
 	/* Check for stdio redirect settings and do the redirection */
 	if ((env_str = SDL_getenv("SDL_STDIO_REDIRECT"))) {
 		if (SDL_atoi(env_str)) {
@@ -408,6 +261,7 @@ int WINAPI WinMain(HINSTANCE hInst, HINSTANCE hPrev, LPSTR szCmdLine, int sw)
 		redirect_output();
 	}
 #endif
+#endif // 0
 
 #ifdef _WIN32_WCE
 	nLen = wcslen(szCmdLine)+128+1;
